Code cleanup + mu vs k script
This change moves all configuration directly to the init method of the
Simulation class. Previously configuration was split between the
Simulation class and the Simulation.perform_simulation method.

Context management was added for better error handling, and so that
killing an experiment mid-run will still allow a gif to be written for
plots created up to that point.

Membrane and fluid classes were moved to their own module and a
"write" option was added to toggle whether plots and datafiles are
written. A warnings property was added to the Simulation to track
whether warnings occur; this may be used to catch overflow and
DivideByZero warnings so that we can either stop if the simulation is
unstable, or log our own warnings.

Finally, a simple script that demonstrates the linear relationship between mu/k
and the min dt that will blow up the simulation.